Implement Rabin-Karp multiple pattern search
This commit adds the necessary code changes required for Rabin-Karp to
search multiple patterns in a string. Bloom Filter is used. The main
changes are:
    - Rabin-Karp now uses Bloom Filter instead of a hash function. Hence
      the hash functions have been removed.
    - The unit tests have been rewritten to test the new changes.
    - The GUI logic responsible for searching a needle inside a haystack
      has been removed for now as it needs work to be compatible with the
      changed function prototype.

Note: Currently the algorithm is limited so that it only searches for
multiple patterns when they are of the same length. Searching for patterns
with varying lengths is, at this point, undefined behaviour.