Add varying pattern length search
The Rabin-Karp now works for patterns with varying lengths. It
accomplishes this by hashing only the first n characters of every pattern,
where n is the length of the smallest pattern. This strategy definitely
increases the risk of hash collisions, but we need more testing to ensure
if it is a problem for us.