Add another hash function in Bloom Filter
Bloom Filter's strength lies in using multiple hash functions to reduce
collisions. In an effort to increase accuracy, a new hash function
"Bernstein hash" now works together with the "Rabin-Karp rolling hash" to
insert and check entries in the Bloom Filter.

With this addition, accuracy has certainly improved. To further minimze
collisions, the size of the Bloom Filter has also been increased to a new
prime number. The size is still under 250 bytes.