Add detection with data files
Plagiarism detection is now automated through the addition of a new
function (detect()). This function scans the ./data directory for the best
matching file with the content in the input control, and reads its content
into the output control. Then the matching words are highlighted in the
input control just like before.